fix(sync): resolve CodeQL blockers — break import cycle, sanitize logging, fix tautological assertion (CHAOS-2581)
- Break dispatch_outbox<->planner import cycle: move build_post_sync_dispatch_payload + PostSyncDispatchPayload to workers/sync_runtime.py (deferred planner import); dispatch_outbox.py is now a pure CAS/claim helper with no planner dependency.
- Sanitize log-injection in admin/routers/integrations.py: constant log message strings + structured extra fields (integration_id, sync_run_id, error, error_type), no raw user value in message.
- Replace tautological no-gap assertion in test_chaos_2581_invariants.py with a real durable-state check (persisted backfill/incremental windows overlap at last_backfill_before; no SyncWatermark rows for the composition).
