+++ src/dev_health_ops/api/admin/routers/integrations.py
@@ -40,6 +40,16 @@ from .common import get_session
 
 logger = logging.getLogger(__name__)
 
+
+def _safe_log_value(value: object, *, max_length: int = 500) -> str:
+    text = str(value)
+    sanitized = "".join(
+        char if char.isprintable() and char not in {"\n", "\r", "\t"} else " "
+        for char in text
+    )
+    return sanitized[:max_length]
+
+
 router = APIRouter()
 
 
@@ -259,7 +269,14 @@ async def discover_integration_sources(
     except ValueError as exc:
         raise HTTPException(status_code=400, detail=str(exc))
     except Exception as exc:
-        logger.exception("Discovery failed for integration %s: %s", integration_id, exc)
+        logger.exception(
+            "integration_discovery.failed",
+            extra={
+                "integration_id": _safe_log_value(integration_id),
+                "error": _safe_log_value(exc),
+                "error_type": type(exc).__name__,
+            },
+        )
         raise HTTPException(status_code=503, detail=f"Discovery failed: {exc}")
 
     return DiscoverResponse(
@@ -422,9 +439,10 @@ async def trigger_integration_sync(
         logger.warning(
             "integration_sync.dispatch_fastpath_failed",
             extra={
-                "integration_id": integration_id,
-                "sync_run_id": plan.sync_run_id,
-                "error": str(exc),
+                "integration_id": _safe_log_value(integration_id),
+                "sync_run_id": _safe_log_value(plan.sync_run_id),
+                "error": _safe_log_value(exc),
+                "error_type": type(exc).__name__,
             },
         )
 
@@ -486,9 +504,10 @@ async def trigger_integration_backfill(
         logger.warning(
             "integration_backfill.dispatch_fastpath_failed",
             extra={
-                "integration_id": integration_id,
-                "sync_run_id": plan.sync_run_id,
-                "error": str(exc),
+                "integration_id": _safe_log_value(integration_id),
+                "sync_run_id": _safe_log_value(plan.sync_run_id),
+                "error": _safe_log_value(exc),
+                "error_type": type(exc).__name__,
             },
         )
